+++ Changes.md
@@ -1,3 +1,5 @@
+* #427 - run the scheduler and table optimizer in a background thread
+
 Release 10.4
 ============
 * #22 - allow one to override the auto-generated finalist schedule